+++ src/react/loom-app/row-options/styles.css
@@ -1,0 +1,10 @@
+.dataloom-row-options {
+    width: 100%;
+    height: 100%;
+}
+
+.dataloom-row-options__container {
+    width: 100%;
+    height: 100%;
+    padding-left: 5px;
+}
